+++ src/briefy/leica/models/__init__.py
@@ -1,5 +1,6 @@
 """Database models to work within the briefy.leica system."""
 from briefy.common.db.models.item import Item
+from briefy.common.db.models.local_role import LocalRole
 from briefy.common.db.models.roles import LocalRoleDeprecated
 from briefy.leica.db import Session
 from briefy.leica.models.asset import Asset
@@ -45,7 +46,9 @@ from briefy.ws.listeners import register_workflow_context_listeners
 import sqlalchemy as sa
 
 
+# force session in these models
 Item.__session__ = Session
+LocalRole.__session__ = Session
 LocalRoleDeprecated.__session__ = Session
 
 ALL_MODELS = [